Created playground and click input
Made a playground to be used for future testing of the grappling hook feature, as well as wrote the rudimentary logic for when the player is grappling or not.

Also altered project settings to make the game more narrow to fit its vertical movement, and maintain its viewport regardless of screen size.

Next intended step is implementation of the grappling hook physics and visuals.